Code review changes

+++ src/RegisterMapping.h
@@ -87,7 +87,7 @@ public:
      * @param deadbandValue indicates a change in value of the register that is insignificant data
      */
     RegisterMapping(const std::string& reference, RegisterType registerType, int32_t address,
-                    bool readRestricted = false, int16_t slaveAddress = -1, float deadbandValue = 0.0);
+                    bool readRestricted = false, int16_t slaveAddress = -1, double deadbandValue = 0.0);
 
     /**
      * @brief Default constructor for mapping with custom OutputType.
@@ -102,7 +102,7 @@ public:
      * @param deadbandValue indicates a change in value of the register that is insignificant data
      */
     RegisterMapping(const std::string& reference, RegisterType registerType, int32_t address, OutputType type,
-                    bool readRestricted = false, int16_t slaveAddress = -1, float deadbandValue = 0.0);
+                    bool readRestricted = false, int16_t slaveAddress = -1, double deadbandValue = 0.0);
 
     /**
      * @brief Constructor for cases where bit is taken from a 16 bit register.
@@ -135,7 +135,7 @@ public:
      */
     RegisterMapping(const std::string& reference, RegisterType registerType, const std::vector<int32_t>& addresses,
                     OutputType type, OperationType operation, bool readRestricted = false, int16_t slaveAddress = -1,
-                    float deadbandValue = 0.0);
+                    double deadbandValue = 0.0);
 
     virtual ~RegisterMapping() = default;
 
@@ -236,7 +236,7 @@ protected:
     std::vector<uint16_t> m_byteValues;
     bool m_isInitialized = false;
     bool m_isValid = false;
-    float m_deadbandValue = 0.0;
+    double m_deadbandValue = 0.0;
 };
 }    // namespace wolkabout
